# feat: Modernize BAM from CoffeeScript to JavaScript with Bun support