… .claude/agents/cli-software-architect.md
… .claude/agents/cli-test-engineer.md
… .claude/agents/cli-tools-developer.md
… .claude/commands/cook.md
… .claude/commands/prp.md
… .github/workflows/build-bun.yml
… CLAUDE.md
… IMPLEMENTATION-SUMMARY.md
… MIGRATION-GUIDE.md
… MIGRATION_ROADMAP.md
… PRPs/bam-modernization-prp.md
… README-ES6.md
… README-MODERNIZED.md
… benchmarks/runner.js
… bun.build.js
… bun/README.md
… bun/build.js
… bun/bun.lock
… bun/cli.js
… bun/commands/gen.js
… bun/commands/new.js
… bun/commands/run.js
… bun/commands/serve.js
… bun/package.json
… bun/utils/assets.js
… bun/utils/page.js
… bunfig.toml
… dist/build-all.sh
… migration/dependency-analysis.md
… migration/test-deps.js